extract FatalErrorBoundary to file
